Migrating product editor images section to slot-fill (#36461)
