+++ plugins/woocommerce-admin/client/products/fills/constants.ts
@@ -1,0 +1,7 @@
+export const PRODUCT_DETAILS_SLUG = 'product-details';
+
+export const DETAILS_SECTION_ID = 'general/details';
+export const IMAGES_SECTION_ID = 'general/images';
+
+export const TAB_GENERAL_ID = 'tab/general';
+export const PLUGIN_ID = 'woocommerce';
