--- plugins/woocommerce-admin/client/products/fills/details-section/constants.ts
@@ -1,2 +1,0 @@
-export const PRODUCT_DETAILS_SLUG = 'product-details';
-export const DETAILS_SECTION_ID = 'general/details';
