+++ plugins/woocommerce-admin/client/products/fills/details-section/index.ts
@@ -3,4 +3,3 @@ export * from './details-field-categories';
 export * from './details-field-feature';
 export * from './details-field-summary';
 export * from './details-field-description';
-export * from './constants';
