+++ plugins/woocommerce-admin/client/products/fills/images-section/images-section.scss
