+++ plugins/woocommerce-admin/client/products/fills/images-section/index.ts
@@ -1,0 +1,1 @@
+export * from './images-field-gallery';
