+++ plugins/woocommerce-admin/client/products/fills/index.ts
@@ -4,3 +4,4 @@
 import './product-form-fills';
 
 export * from './details-section/details-section-fills';
+export * from './images-section/images-section-fills';
