+++ plugins/woocommerce-admin/client/products/product-form.tsx
@@ -20,12 +20,12 @@ import { ProductInventorySection } from './sections/product-inventory-section';
 import { PricingSection } from './sections/pricing-section';
 import { ProductShippingSection } from './sections/product-shipping-section';
 import { ProductVariationsSection } from './sections/product-variations-section';
-import { ImagesSection } from './sections/images-section';
 import { validate } from './product-validation';
 import { AttributesSection } from './sections/attributes-section';
 import { OptionsSection } from './sections/options-section';
 import { ProductFormFooter } from './layout/product-form-footer';
 import { ProductFormTab } from './product-form-tab';
+import { TAB_GENERAL_ID } from './fills/constants';
 
 export const ProductForm: React.FC< {
 	product?: PartialProduct;
@@ -50,8 +50,9 @@ export const ProductForm: React.FC< {
 				<ProductFormHeader />
 				<ProductFormLayout>
 					<ProductFormTab name="general" title="General">
-						<WooProductSectionItem.Slot location="tab/general" />
-						<ImagesSection />
+						<WooProductSectionItem.Slot
+							location={ TAB_GENERAL_ID }
+						/>
 						<AttributesSection />
 					</ProductFormTab>
 					<ProductFormTab
