+++ plugins/woocommerce/changelog/add-36417-mvp-images-slotfill
@@ -1,0 +1,4 @@
+Significance: minor
+Type: add
+
+Using slotfill to insert images section in product editor.
